Fix imports to avoid import errors
