+++ saspy/__init__.py
@@ -19,7 +19,9 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 from saspy.version import __version__
-from saspy.sasbase import SASdata, SASsession, SASconfig
+from saspy.sasbase import SASsession, SASconfig
+# from saspy.sasdata import SASdata
+# from saspy.sasproccommons import SASProcCommons
 
 def isnotebook():
     try:
