+++ saspy/sasqc.py
@@ -15,10 +15,12 @@
 #
 
 import logging
+from typing import TYPE_CHECKING
+from saspy.sasproccommons import procDecorator
 
-from saspy.sasdecorator import procDecorator
-from saspy.sasresults import SASresults
-#from saspy.sasbase import SASdata
+if TYPE_CHECKING:
+    from saspy.sasresults import SASresults
+    from saspy.sasdata import SASdata
 
 
 class SASqc:
